Fix border generation, padding
Use passed in border values when bordering individual images in the stack
Fix generation of image padding, and try to streamline this. I don't
remember the reason for wrapping the borderImage call and composing it. This
seems to work locally, though I need to test it with older Im versions.